Added path for sanity.types.ts & Updated sanity.types.ts
