+++ types/sanity.types.ts
@@ -140,6 +140,92 @@ export type ColorTheme = {
     background?: Color;
 };
 
+export type MenuItem = {
+    _type: "menuItem";
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    linkInternal?: string;
+};
+
+export type Stage = {
+    _type: "stage";
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    description?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+};
+
+export type Customer = {
+    _type: "customer";
+    name?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    position?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    photo?: {
+        asset?: {
+            _ref: string;
+            _type: "reference";
+            _weak?: boolean;
+            [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
+        };
+        hotspot?: SanityImageHotspot;
+        crop?: SanityImageCrop;
+        _type: "image";
+    };
+};
+
+export type AnchorLink = {
+    _type: "anchorLink";
+    subtitle?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    anchorId?: Slug;
+};
+
+export type AboutUsItem = {
+    _type: "aboutUsItem";
+    AboutUs?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+};
+
+export type SocialLinks = {
+    _type: "socialLinks";
+    platform?:
+        | "linkedin"
+        | "telegram"
+        | "github"
+        | "behance"
+        | "googleDrive"
+        | "whatsapp"
+        | "youtube"
+        | "pinterest"
+        | "email"
+        | "facebook"
+        | "instagram"
+        | "twitter";
+    url?: LinkExternal;
+};
+
 export type Milestone = {
     _type: "milestone";
     title?: string;
@@ -178,15 +264,6 @@ export type Timeline = {
         _key: string;
     }>;
 };
-
-export type MenuLinks = Array<
-    | ({
-          _key: string;
-      } & LinkInternal)
-    | ({
-          _key: string;
-      } & LinkExternal)
->;
 
 export type Instagram = {
     _type: "instagram";
@@ -292,58 +369,6 @@ export type GridItem = {
     }>;
 };
 
-// export type CallToAction = {
-//     _type: "callToAction";
-//     layout?: "left" | "right";
-//     title?: string;
-//     portableText?: string;
-//     link?: Array<
-//         | ({
-//               _key: string;
-//           } & LinkInternal)
-//         | ({
-//               _key: string;
-//           } & LinkExternal)
-//     >;
-//     content?: Array<{
-//         asset?: {
-//             _ref: string;
-//             _type: "reference";
-//             _weak?: boolean;
-//             [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
-//         };
-//         hotspot?: SanityImageHotspot;
-//         crop?: SanityImageCrop;
-//         _type: "image";
-//         _key: string;
-//     }>;
-// };
-export type InternationalizedArrayPortableTextSimpleValue = {
-    _type: "internationalizedArrayPortableTextSimpleValue";
-    value?: PortableTextSimple;
-};
-
-export type InternationalizedArrayPortableTextSimple = Array<
-    {
-        _key: string;
-    } & InternationalizedArrayPortableTextSimpleValue
->;
-
-export type Button = {
-    _type: "button";
-    buttonName?: InternationalizedArrayString;
-    buttonLink?: "noLink" | "internal" | "external";
-    linkInternal?: LinkInternal;
-    linkExternal?: LinkExternal;
-};
-
-export type CallToAction = {
-    _type: "callToAction";
-    title?: InternationalizedArrayString;
-    ctaButton?: Button;
-    description?: InternationalizedArrayPortableTextSimple;
-};
-
 export type Callout = {
     _type: "callout";
     text?: string;
@@ -402,6 +427,21 @@ export type AccordionGroup = {
         _key: string;
     }>;
 };
+
+export type PortableColorTitle = Array<{
+    children?: Array<{
+        marks?: Array<string>;
+        text?: string;
+        _type: "span";
+        _key: string;
+    }>;
+    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
+    listItem?: never;
+    markDefs?: null;
+    level?: number;
+    _type: "block";
+    _key: string;
+}>;
 
 export type PortableTextSimple = Array<{
     children?: Array<{
@@ -478,29 +518,452 @@ export type PortableText = Array<
       } & Instagram)
 >;
 
-export type PortableColorTitle = Array<{
-    children?: Array<{
-        marks?: Array<"purple" | string>;
-        text?: string;
-        _type: "span";
-        _key: string;
-    }>;
-    style?: "normal";
-    markDefs?: Array<{
-        _key: string;
-    }>;
-    _type: "block";
-    _key: string;
-}>;
+export type Tool = {
+    _id: string;
+    _type: "tool";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    title?: string;
+};
 
-export type LinkInternal = {
-    _type: "linkInternal";
-    reference?: {
+export type Service = {
+    _id: string;
+    _type: "service";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    description?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+};
+
+export type Faq = {
+    _id: string;
+    _type: "faq";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    numberOfLikes?: number;
+    question?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    shortAnswer?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    fullAnswer?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableTextValue
+    >;
+    image?: {
+        asset?: {
+            _ref: string;
+            _type: "reference";
+            _weak?: boolean;
+            [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
+        };
+        hotspot?: SanityImageHotspot;
+        crop?: SanityImageCrop;
+        caption?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayStringValue
+        >;
+        _type: "image";
+    };
+};
+
+export type Review = {
+    _id: string;
+    _type: "review";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    project?: {
         _ref: string;
         _type: "reference";
         _weak?: boolean;
-        [internalGroqTypeReferenceTo]?: "home";
+        [internalGroqTypeReferenceTo]?: "project";
     };
+    reviewText?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    reviewUrl?: LinkExternal;
+};
+
+export type Project = {
+    _id: string;
+    _type: "project";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    description?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    customers?: Array<
+        {
+            _key: string;
+        } & Customer
+    >;
+    startDate?: string;
+    endDate?: string;
+    category?: {
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        [internalGroqTypeReferenceTo]?: "siteCategory";
+    };
+    status?: "recruitingTeam" | "development" | "completed";
+    teamMembers?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "team";
+    }>;
+    URL?: LinkExternal;
+    image?: {
+        asset?: {
+            _ref: string;
+            _type: "reference";
+            _weak?: boolean;
+            [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
+        };
+        hotspot?: SanityImageHotspot;
+        crop?: SanityImageCrop;
+        caption?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayStringValue
+        >;
+        _type: "image";
+    };
+    stages?: Array<string>;
+};
+
+export type SiteCategory = {
+    _id: string;
+    _type: "siteCategory";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    categoryName?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+};
+
+export type Team = {
+    _id: string;
+    _type: "team";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    name?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    role?: {
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        [internalGroqTypeReferenceTo]?: "specialist";
+    };
+    price?: number;
+    isAvaliblePerson?: "Show price" | "Do not show price";
+    about?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    services?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    projects?: Array<
+        | ({
+              _key: string;
+          } & LinkExternal)
+        | {
+              _ref: string;
+              _type: "reference";
+              _weak?: boolean;
+              _key: string;
+              [internalGroqTypeReferenceTo]?: "project";
+          }
+    >;
+    photo?: {
+        asset?: {
+            _ref: string;
+            _type: "reference";
+            _weak?: boolean;
+            [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
+        };
+        hotspot?: SanityImageHotspot;
+        crop?: SanityImageCrop;
+        _type: "image";
+    };
+    socialLinks?: Array<
+        {
+            _key: string;
+        } & SocialLinks
+    >;
+    tools?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "tool";
+    }>;
+};
+
+export type Specialist = {
+    _id: string;
+    _type: "specialist";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    title?: string;
+    category?: {
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        [internalGroqTypeReferenceTo]?: "specialistCategory";
+    };
+};
+
+export type SpecialistCategory = {
+    _id: string;
+    _type: "specialistCategory";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+};
+
+export type LinkInternal = {
+    _type: "linkInternal";
+    reference?:
+        | {
+              _ref: string;
+              _type: "reference";
+              _weak?: boolean;
+              [internalGroqTypeReferenceTo]?: "home";
+          }
+        | {
+              _ref: string;
+              _type: "reference";
+              _weak?: boolean;
+              [internalGroqTypeReferenceTo]?: "aboutUs";
+          };
+};
+
+export type AboutUs = {
+    _id: string;
+    _type: "aboutUs";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    language?: string;
+};
+
+export type Home = {
+    _id: string;
+    _type: "home";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    hero?: HeroHome;
+    aboutUsHomeSection?: AboutUsHomeSection;
+    ctaSectionJoinUs?: CallToAction;
+    portfolioHome?: PortfolioHome;
+    reviewsHome?: ReviewsHome;
+    ctaSectionWriteUs?: CallToAction;
+    servicesHome?: ServicesHome;
+    teamHome?: TeamHome;
+    ctaSectionJoinUsSecond?: CallToAction;
+    stagesHome?: StagesHome;
+    faqHome?: FaqHome;
+    ctaSectionOrder?: CallToAction;
+};
+
+export type FaqHome = {
+    _type: "faqHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    faqList?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "faq";
+    }>;
+};
+
+export type StagesHome = {
+    _type: "stagesHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    stagesList?: Array<
+        {
+            _key: string;
+        } & Stage
+    >;
+};
+
+export type TeamHome = {
+    _type: "teamHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    projectsList?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "project";
+    }>;
+};
+
+export type ServicesHome = {
+    _type: "servicesHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    servicesList?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "service";
+    }>;
+    description?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+};
+
+export type ReviewsHome = {
+    _type: "reviewsHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    reviewsSlider?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "review";
+    }>;
+};
+
+export type PortfolioHome = {
+    _type: "portfolioHome";
+    sectionId?: AnchorLink;
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    portfolioSlider?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "project";
+    }>;
+};
+
+export type CallToAction = {
+    _type: "callToAction";
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayStringValue
+    >;
+    description?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableTextSimpleValue
+    >;
+};
+
+export type AboutUsHomeSection = {
+    _type: "aboutUsHomeSection";
+    sectionId?: AnchorLink;
+    sectionTitle?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayPortableColorTitleValue
+    >;
+    shortDescription?: Array<
+        {
+            _key: string;
+        } & AboutUsItem
+    >;
+    learnMoreButton?: Button;
+};
+
+export type HeroHome = {
+    _type: "heroHome";
+    title?: Array<
+        {
+            _key: string;
+        } & InternationalizedArrayTextValue
+    >;
+    portfolioSlider?: Array<{
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        _key: string;
+        [internalGroqTypeReferenceTo]?: "project";
+    }>;
 };
 
 export type LinkExternal = {
@@ -514,10 +977,54 @@ export type LinkEmail = {
     email?: string;
 };
 
+export type InternationalizedArrayPortableColorTitleValue = {
+    _type: "internationalizedArrayPortableColorTitleValue";
+    value?: PortableColorTitle;
+};
+
+export type InternationalizedArrayPortableTextSimpleValue = {
+    _type: "internationalizedArrayPortableTextSimpleValue";
+    value?: PortableTextSimple;
+};
+
+export type InternationalizedArrayPortableTextValue = {
+    _type: "internationalizedArrayPortableTextValue";
+    value?: PortableText;
+};
+
+export type InternationalizedArrayTextValue = {
+    _type: "internationalizedArrayTextValue";
+    value?: string;
+};
+
 export type InternationalizedArrayStringValue = {
     _type: "internationalizedArrayStringValue";
     value?: string;
 };
+
+export type InternationalizedArrayPortableColorTitle = Array<
+    {
+        _key: string;
+    } & InternationalizedArrayPortableColorTitleValue
+>;
+
+export type InternationalizedArrayPortableTextSimple = Array<
+    {
+        _key: string;
+    } & InternationalizedArrayPortableTextSimpleValue
+>;
+
+export type InternationalizedArrayPortableText = Array<
+    {
+        _key: string;
+    } & InternationalizedArrayPortableTextValue
+>;
+
+export type InternationalizedArrayText = Array<
+    {
+        _key: string;
+    } & InternationalizedArrayTextValue
+>;
 
 export type InternationalizedArrayString = Array<
     {
@@ -541,19 +1048,12 @@ export type TranslationMetadata = {
 
 export type InternationalizedArrayReferenceValue = {
     _type: "internationalizedArrayReferenceValue";
-    value?:
-        | {
-              _ref: string;
-              _type: "reference";
-              _weak?: boolean;
-              [internalGroqTypeReferenceTo]?: "home";
-          }
-        | {
-              _ref: string;
-              _type: "reference";
-              _weak?: boolean;
-              [internalGroqTypeReferenceTo]?: "settings";
-          };
+    value?: {
+        _ref: string;
+        _type: "reference";
+        _weak?: boolean;
+        [internalGroqTypeReferenceTo]?: "settings";
+    };
 };
 
 export type Settings = {
@@ -563,48 +1063,21 @@ export type Settings = {
     _updatedAt: string;
     _rev: string;
     language?: string;
-    menu?: Menu;
-    footer?: FooterSettings;
+    header?: Header;
+    footer?: Footer;
     notFoundPage?: NotFoundPage;
     seo?: Seo;
+    buttonBuyMeCoffee?: Button;
+    buttonJoinUS?: Button;
+    buttonOrder?: Button;
 };
 
-export type NotFoundPage = {
-    _type: "notFoundPage";
-    title?: string;
-    body?: string;
-};
-
-export type FooterSettings = {
-    _type: "footerSettings";
-    links?: Array<
-        | ({
-              _key: string;
-          } & LinkInternal)
-        | ({
-              _key: string;
-          } & LinkExternal)
-    >;
-    text?: PortableTextSimple;
-};
-
-export type Menu = {
-    _type: "menu";
-    links?: MenuLinks;
-};
-
-export type Home = {
-    _id: string;
-    _type: "home";
-    _createdAt: string;
-    _updatedAt: string;
-    _rev: string;
-    language?: string;
-    hero?: Hero;
-    seo?: Seo;
-    ctaSectionJoinUs?: CallToAction;
-    ctaSectionOrder?: CallToAction;
-    ctaSectionWriteUs?: CallToAction;
+export type Button = {
+    _type: "button";
+    buttonName?: InternationalizedArrayString;
+    buttonLink?: "noLink" | "internal" | "external";
+    linkInternal?: LinkInternal;
+    linkExternal?: LinkExternal;
 };
 
 export type SanityImageCrop = {
@@ -666,7 +1139,7 @@ export type SanityImageMetadata = {
 
 export type Seo = {
     _type: "seo";
-    title?: string;
+    title?: InternationalizedArrayString;
     description?: string;
     image?: {
         asset?: {
@@ -681,17 +1154,45 @@ export type Seo = {
     };
 };
 
-export type Hero = {
-    _type: "hero";
-    title?: string;
-    description?: string;
-    link?: Array<
-        | ({
-              _key: string;
-          } & LinkInternal)
-        | ({
-              _key: string;
-          } & LinkExternal)
+export type NotFoundPage = {
+    _type: "notFoundPage";
+    title?: InternationalizedArrayString;
+    description?: InternationalizedArrayText;
+    goToHomeButton?: Button;
+};
+
+export type Footer = {
+    _type: "footer";
+    title?: InternationalizedArrayPortableColorTitle;
+    navigationMenu?: Array<
+        {
+            _key: string;
+        } & MenuItem
+    >;
+    privacyPolicy?: {
+        title?: InternationalizedArrayString;
+        url?: Array<
+            | ({
+                  _key: string;
+              } & LinkExternal)
+            | ({
+                  _key: string;
+              } & LinkInternal)
+        >;
+    };
+};
+
+export type Header = {
+    _type: "header";
+    navigationMenu?: Array<
+        {
+            _key: string;
+        } & MenuItem
+    >;
+    socialLinks?: Array<
+        {
+            _key: string;
+        } & SocialLinks
     >;
 };
 
@@ -757,42 +1258,73 @@ export type AllSanitySchemaTypes =
     | Geopoint
     | Page
     | ColorTheme
+    | MenuItem
+    | Stage
+    | Customer
+    | AnchorLink
+    | AboutUsItem
+    | SocialLinks
     | Milestone
     | Duration
     | Timeline
-    | MenuLinks
     | Instagram
     | ImageFeature
     | Images
     | ImageCallToAction
     | Grid
     | GridItem
-    | CallToAction
     | Callout
     | Accordion
     | AccordionGroup
+    | PortableColorTitle
     | PortableTextSimple
     | PortableText
-    | PortableColorTitle
+    | Tool
+    | Service
+    | Faq
+    | Review
+    | Project
+    | SiteCategory
+    | Team
+    | Specialist
+    | SpecialistCategory
     | LinkInternal
+    | AboutUs
+    | Home
+    | FaqHome
+    | StagesHome
+    | TeamHome
+    | ServicesHome
+    | ReviewsHome
+    | PortfolioHome
+    | CallToAction
+    | AboutUsHomeSection
+    | HeroHome
     | LinkExternal
     | LinkEmail
+    | InternationalizedArrayPortableColorTitleValue
+    | InternationalizedArrayPortableTextSimpleValue
+    | InternationalizedArrayPortableTextValue
+    | InternationalizedArrayTextValue
     | InternationalizedArrayStringValue
+    | InternationalizedArrayPortableColorTitle
+    | InternationalizedArrayPortableTextSimple
+    | InternationalizedArrayPortableText
+    | InternationalizedArrayText
     | InternationalizedArrayString
     | TranslationMetadata
     | InternationalizedArrayReferenceValue
     | Settings
-    | NotFoundPage
-    | FooterSettings
-    | Menu
-    | Home
+    | Button
     | SanityImageCrop
     | SanityImageHotspot
     | SanityImageAsset
     | SanityAssetSourceData
     | SanityImageMetadata
     | Seo
-    | Hero
+    | NotFoundPage
+    | Footer
+    | Header
     | InternationalizedArrayReference
     | Color
     | RgbaColor
@@ -803,17 +1335,75 @@ export type AllSanitySchemaTypes =
 export declare const internalGroqTypeReferenceTo: unique symbol;
 // Source: ./sanity/lib/queries.ts
 // Variable: homePageQuery
-// Query:   *[_type == "home"][0]{    _id,    title,    hero,    language  }
+// Query: *[_type == "home"][0]{  ...,  hero {  ...,    "title": title[_key == $language][0].value  },  ctaSectionJoinUs {  ...,    "ctaButton": ctaButton {"buttonName": buttonName[_key == $language][0].value}  },  ctaSectionOrder {  ...,    "ctaButton": ctaButton {"buttonName": buttonName[_key == $language][0].value}  }, }
 export type HomePageQueryResult = {
     _id: string;
-    title: null;
-    hero: Hero | null;
-    language: string | null;
+    _type: "home";
+    _createdAt: string;
+    _updatedAt: string;
+    _rev: string;
+    hero: {
+        _type: "heroHome";
+        title: string | null;
+        portfolioSlider?: Array<{
+            _ref: string;
+            _type: "reference";
+            _weak?: boolean;
+            _key: string;
+            [internalGroqTypeReferenceTo]?: "project";
+        }>;
+    } | null;
+    aboutUsHomeSection?: AboutUsHomeSection;
+    ctaSectionJoinUs: {
+        _type: "callToAction";
+        title?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayStringValue
+        >;
+        description?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayPortableTextSimpleValue
+        >;
+        ctaButton: null;
+    } | null;
+    portfolioHome?: PortfolioHome;
+    reviewsHome?: ReviewsHome;
+    ctaSectionWriteUs?: CallToAction;
+    servicesHome?: ServicesHome;
+    teamHome?: TeamHome;
+    ctaSectionJoinUsSecond?: CallToAction;
+    stagesHome?: StagesHome;
+    faqHome?: FaqHome;
+    ctaSectionOrder: {
+        _type: "callToAction";
+        title?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayStringValue
+        >;
+        description?: Array<
+            {
+                _key: string;
+            } & InternationalizedArrayPortableTextSimpleValue
+        >;
+        ctaButton: null;
+    } | null;
 } | null;
 // Variable: settingsQuery
-// Query:   *[_type == "settings"][0]{    footer,    menuItems[]->{      _type,      "slug": slug.current,      title    },    ogImage,  }
+// Query: *[_type == "settings"][0]{    footer,    menuItems[]->{      _type,      "slug": slug.current,      title    },    ogImage,  }
 export type SettingsQueryResult = {
-    footer: FooterSettings | null;
+    footer: Footer | null;
     menuItems: null;
     ogImage: null;
 } | null;
+
+// Query TypeMap
+import "@sanity/client";
+declare module "@sanity/client" {
+    interface SanityQueries {
+        '\n  *[_type == "home"][0]{\n  ...,\n  hero {\n  ...,\n    "title": title[_key == $language][0].value\n  },\n  ctaSectionJoinUs {\n  ...,\n    "ctaButton": ctaButton {"buttonName": buttonName[_key == $language][0].value}\n  },\n  ctaSectionOrder {\n  ...,\n    "ctaButton": ctaButton {"buttonName": buttonName[_key == $language][0].value}\n  },\n \n}': HomePageQueryResult;
+        '\n  *[_type == "settings"][0]{\n    footer,\n    menuItems[]->{\n      _type,\n      "slug": slug.current,\n      title\n    },\n    ogImage,\n  }\n': SettingsQueryResult;
+    }
+}
